ref: hooks folder restructurized
