+++ src/components/map/map.tsx
@@ -4,8 +4,9 @@ import 'leaflet/dist/leaflet.css';
 import { memo, useEffect, useRef } from 'react';
 import { useLocation } from 'react-router-dom';
 import { AppRoute, URL_MARKER_CURRENT, URL_MARKER_DEFAULT } from '../../const';
-import { useMap } from '../../hooks/use-map';
+
 import { LocationType, OfferType, PlaceCardType } from '../../types';
+import { useMap } from '../../hooks/use-map/use-map';
 
 // ^======================== map ========================^ //
 type MapProps = {
